refactor: route 4 components through facades instead of direct service imports
- config-settings-panel.js and settings-configs.js now use new configFacade
- webview-panel.js now uses existing boardFacade.openExternal
- git-changes-view.js now uses extended tabViewFacade (gitLocalChanges, gitFileDiff)
